feat: array support for msg in cli
