+++ packages/cli/src/react/jsx/utils/stringParsing/types.ts
@@ -7,8 +7,17 @@ import { Updates } from '../../../../types/index.js';
 export type ParsingConfig = {
   parsingOptions: ParsingConfigOptions;
   file: string;
-  ignoreAdditionalData: boolean;
+  /**
+   * If true, ignores fields like $context, $id, etc. when extracting metadata from a string entry
+   */
+  ignoreInlineMetadata: boolean;
+  /**
+   * If true, dynamic content will not be treated as errors
+   */
   ignoreDynamicContent: boolean;
+  /**
+   * If true, invalid ICU strings will not be treated as errors
+   */
   ignoreInvalidIcu: boolean;
 };
